requirements.txt
PyQt5>=5.15
PySide6>= 5.15
grpcio-tools
